+++ R/calc_cor.R
@@ -64,6 +64,8 @@ setMethod(
 #'   produced by [circlize::colorRamp2()].
 #' @param ... other arguments to be passed to [ComplexHeatmap::Heatmap()].
 #'
+#' @importFrom ComplexHeatmap rowAnnotation HeatmapAnnotation Heatmap
+#' @importFrom circlize colorRamp2
 #' @export
 #'
 #' @examples