+++ R/pca_cor_samplevar.R
@@ -169,6 +169,10 @@ setMethod(
 #'   produced by [circlize::colorRamp2()].
 #' @param ... other arguments to be passed to [ComplexHeatmap::Heatmap()].
 #'
+#' @importFrom ComplexHeatmap Heatmap
+#' @importFrom circlize colorRamp2
+#' @export
+#'
 #' @examples
 #' autoplot(result)
 #'